A.) Phycsics: Lets make the game as fluid and spaced like as possible with the way the ship moves through the envoirment. Lets make the ship move just like this ship in this game... Last Horizons on the app store. B.) The Environment. I really want the background of the environment to move in a moving 2d game. Let's get the mood going for how the game will playout. Just like Full of Stars on the app store. If you cant find anything in the files we have provided. Please don't hesitate to go to a asset store online and find what you want and send to us for approval. C.) Track Size: The first level should take around 15-20 seconds to circle around. D.) Lets add in the faster paced music for the level. It has been provide in the asset folder. E.) Let's have a 3...2...1 count down to start the level.
A.) Lock the screen so that you cannot turn it horizontally. It makes it hard to play for now. (You may have to make the ship a little smaller and position it a tab bit lower on the screen so that it is playable for higher speed levels. Basically taking the aspect scale of the horizontal play mode and making it that size for the vertical aspect scale.            ////finish
B.) Make sure the lasers shoot as far as the screen goes. Right now they stop after a few shots in.
C.) Lets put life to the asteroids. 3 Levels work of damage it takes to destroy them. Small ones easy to explode. Medium harder to explode. ones really have to shoot them alot to make them explode.                                        ///finish
D.) When the ship turns, no need for it to animate turning. Just leave it stationary.                      ///finish
E.) Let's make the score in the top left corner less bricky. When I say that I mean make it move more fluidly between numbers.   ///finish